move toolbar stuff from gallery page
